// src/lib/relationship-types.ts
export type SavedEntry = {
  familyIds: string[];
  brandIds: string[];
  modelIds: string[];
  itemBrandId: string;
  itemPartId: string;
  itemPartModelIds: string[];
  familyNames: string[];
  brandNames: string[];
  modelNames: string[];
  itemBrandName: string;
  itemPartName: string;
  itemPartModelNames: string[];
};